+++ src/Component/Event.php
@@ -101,7 +101,7 @@ class Event extends Component
     protected $organizer;
 
     /**
-     * @see http://www.ietf.org/rfc/rfc2445.txt 4.8.2.7 Time Transparency
+     * @see https://tools.ietf.org/html/rfc5545#section-3.8.2.7
      *
      * @var string
      */