+++ src/Property/StringValue.php
@@ -27,16 +27,24 @@ class StringValue implements ValueInterface
         $this->value = $value;
     }
 
-    /**
-     * Return the value of the Property as an escaped string.
-     *
-     * Escape values as per RFC 2445. See http://www.kanzaki.com/docs/ical/text.html
-     *
-     * @return string
-     */
-    public function getEscapedValue()
+    public function getEscapedValue(): string
     {
-        return PropertyValueUtil::escapeValue((string) $this->value);
+        $value = $this->value;
+
+        $value = str_replace('\\', '\\\\', $value);
+        $value = str_replace('"', '\\"', $value);
+        $value = str_replace(',', '\\,', $value);
+        $value = str_replace(';', '\\;', $value);
+        $value = str_replace("\n", '\\n', $value);
+        $value = str_replace([
+            "\x00", "\x01", "\x02", "\x03", "\x04", "\x05", "\x06", "\x07",
+            "\x08", "\x09", /* \n*/ "\x0B", "\x0C", "\x0D", "\x0E", "\x0F",
+            "\x10", "\x11", "\x12", "\x13", "\x14", "\x15", "\x16", "\x17",
+            "\x18", "\x19", "\x1A", "\x1B", "\x1C", "\x1D", "\x1E", "\x1F",
+            "\x7F",
+        ], '', $value);
+
+        return $value;
     }
 
     /**